--- Product/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-
-server.port=9095
-spring.h2.console.enabled=true
