+++ dockercomposer/docker-compose.yml
@@ -1,5 +1,6 @@
 version: '3'
-services:      
+services:  
+    
   productweb:
     container_name: product-service
     image: product-service
@@ -11,7 +12,8 @@ services:
       ZIPKIN_HOST: zipkin
       ZIPKIN_PORT: 9411   
     depends_on:
-      - eureka      
+      - eureka
+      - mysqldb      
     restart: always  
   
   reviewweb:
@@ -65,3 +67,14 @@ services:
       - productweb
       - eureka
     restart: always
+    
+  mysqldb:
+    container_name: mysqldb
+    hostname: mysqldb
+    image: mysql:5.7
+    environment:
+      MYSQL_DATABASE: product_info
+      MYSQL_PASSWORD: root
+      MYSQL_ROOT_PASSWORD: root
+    ports:
+      - '3306:3306'
