+++ eureka/.gitignore
@@ -27,3 +27,7 @@ HELP.md
 
 ### VS Code ###
 .vscode/
+
+eureka/.mvn/
+eureka/mvnw
+eureka/mvnw.cmd
