+++ product/src/main/java/com/epam/docker/course/model/Product.java
@@ -1,0 +1,69 @@
+package com.epam.docker.course.model;
+
+import java.util.List;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import javax.persistence.Transient;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@Entity
+@Table(name ="Product")
+@JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+public class Product {
+	
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private long prodId;
+	private String name;
+	private String prodInfo;
+	@Transient
+	private List<ProdReviews> prodReviews;	
+
+	public Product() {
+	}
+
+	public Product(long prodId, String name, String prodInfo) {
+		super();
+		this.prodId = prodId;
+		this.name = name;
+		this.prodInfo = prodInfo;
+	}
+
+	public long getProdId() {
+		return prodId;
+	}
+
+	public void setProdId(long prodId) {
+		this.prodId = prodId;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public String getProdInfo() {
+		return prodInfo;
+	}
+
+	public void setProdInfo(String prodInfo) {
+		this.prodInfo = prodInfo;
+	}
+
+	public List<ProdReviews> getProdReviews() {
+		return prodReviews;
+	}
+
+	public void setProdReviews(List<ProdReviews> prodReviews) {
+		this.prodReviews = prodReviews;
+	}
+
+}
