+++ product/src/main/resources/application.properties
@@ -1,0 +1,19 @@
+
+## Spring DATASOURCE (DataSourceAutoConfiguration & DataSourceProperties)
+spring.datasource.url = jdbc:mysql://mysqldb:3306/product_info?useSSL=false
+spring.datasource.username = root
+spring.datasource.password = root
+
+## Hibernate Properties
+# The SQL dialect makes Hibernate generate better SQL for the chosen database
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
+
+# Hibernate ddl auto (create, create-drop, validate, update)
+spring.jpa.hibernate.ddl-auto = create-drop
+spring.jpa.generate-ddl = true
+spring.jpa.show-sql=true
+
+
+
+
+
