+++ product/src/test/java/com/epam/docker/course/ProductApplicationTests.java
