+++ zipkin/.gitignore
@@ -27,3 +27,8 @@ HELP.md
 
 ### VS Code ###
 .vscode/
+
+
+zipkin/.mvn/
+zipkin/mvnw
+zipkin/mvnw.cmd
